+++ layers/GatoGraphQLForWP/plugins/gatographql/CHANGELOG.md
@@ -16,6 +16,7 @@ Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) princip
   - [PRO] Translate posts for Polylang (Gutenberg) (#2667)
   - [PRO] Translate posts for Polylang (Classic editor) (#2667)
   - [PRO] Sync featured image for Polylang (#2669)
+  - [PRO] Sync tags and categories for Polylang (#2670)
 
 ### Fixed
 
@@ -28,6 +29,7 @@ Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) princip
   - Polylang: When publishing a post, translate it to all languages (Gutenberg)
   - Polylang: When publishing a post, translate it to all languages (Classic editor)
   - Polylang: When publishing a post, set the featured image for each language on all translation posts
+  - Polylang: When publishing a post, set the tags and categories for each language on all translation posts
 
 ## 2.2.3 - 26/04/2024